Added datalist support for enabling specific dates.
If datetimepicker finds a "list" tag attached to the date input element,
it parses the datalist options and builds a list of allowed dates. Any
date outside of those is disabled on the day picker. This is useful for
datasets where only specific dates contain data.

The date format of the datalist options is the same one used by datetimepicker,
that is, the one set via the 'format' option or the 'data-format' tag.