Fix bug where arrow is on opposite side
Even using the old defaults, 'bottom-right' css is being applied to a datepicker which is created on the left hand side and the arrow is on the left.

Swapping the CSS moves the arrow to the right side even though the name would now be 'bottom-left'. Moveover, the picker would be on the left, but the arrow on the right of the picker.

This should keep this working as is instead of swapping css. I worry changing the CSS would be a breaking change for many people using this library. Even if it seems correct.